make exec.fifo can be safety read on readonly tmpfs
